+++ test/ppx_test/ppx_sedlex_test.ml
@@ -12,11 +12,11 @@ let clear_tables () =
   Hashtbl.clear P.partitions;
   Hashtbl.clear P.tables
 
-let expand ~ctxt:_ expr =
+let expand_with ~shortest ~ctxt:_ expr =
   reset_state ();
   let loc = Location.none in
   let code_expr, auto =
-    P.handle_sedlex_match ~env:P.builtin_regexps ~map_rhs:Fun.id expr
+    P.handle_sedlex_match ~shortest ~env:P.builtin_regexps ~map_rhs:Fun.id expr
   in
   let code_str = Pprintast.string_of_expression code_expr in
   let dot_str = S.dfa_to_dot auto in
@@ -31,6 +31,12 @@ let expand ~ctxt:_ expr =
 let ext =
   Extension.V3.declare "sedlex_test" Extension.Context.expression
     Ast_pattern.(single_expr_payload __)
-    expand
+    (expand_with ~shortest:false)
 
-let () = Driver.register_transformation "sedlex_test" ~extensions:[ext]
+let ext_shortest =
+  Extension.V3.declare "sedlex_test_shortest" Extension.Context.expression
+    Ast_pattern.(single_expr_payload __)
+    (expand_with ~shortest:true)
+
+let () =
+  Driver.register_transformation "sedlex_test" ~extensions:[ext; ext_shortest]